common-apps: conditional move mans in scons

+++ helper/common-apps-functions-source.sh
@@ -6552,9 +6552,12 @@ function build_scons()
         \
         --optimize=1 \
 
-      mkdir -pv "${INSTALL_FOLDER_PATH}/share/man/man1"
-      mv -v "${INSTALL_FOLDER_PATH}/man/man1"/* "${INSTALL_FOLDER_PATH}/share/man/man1"
-      rm -rv "${INSTALL_FOLDER_PATH}/man/man1"
+      if [ -d "${INSTALL_FOLDER_PATH}/man/man1" ]
+      then
+        mkdir -pv "${INSTALL_FOLDER_PATH}/share/man/man1"
+        mv -v "${INSTALL_FOLDER_PATH}/man/man1"/* "${INSTALL_FOLDER_PATH}/share/man/man1"
+        rm -rv "${INSTALL_FOLDER_PATH}/man/man1"
+      fi
 
     ) 2>&1 | tee "${LOGS_FOLDER_PATH}/${scons_folder_name}/install-output.txt"
